+++ lib/polars/lazy_frame.rb
@@ -45,8 +45,7 @@ module Polars
     #
     # @param source [Object]
     #   Path to a file or a file-like object (by file-like object, we refer to
-    #   objects that have a `read` method, such as a file handler (e.g.
-    #   via builtin `open` function) or `StringIO`).
+    #   objects that have a `read` method, such as a file handler or `StringIO`).
     #
     # @return [LazyFrame]
     #